backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
django-oauth-toolkit==2.3.0
psycopg==3.1.18
djangorestframework-simplejwt==5.3.1